Reusable <DocumentsTable /> component
